* MySQL creates temporary columns for numbers when running COUNT as unsigned integers... which are longs in Java. Reflect this in the code. TODO: Find a way to convert to int in MySQL?

+++ src/server/PacketHandler.java
@@ -267,7 +267,7 @@ public abstract class PacketHandler
         }
         
         // Check if the room has space for the player - this might be unnecessary if the server has Room objects containing the necessary data
-        int cur = (Integer)results.get(0).get("current");
+        long cur = (Long)results.get(0).get("current");
         int max = (Integer)results.get(0).get("max_players");
         if((max - cur) == 0)
         {
